fix gmdate

+++ Source/GridDominance.Server/admin/entrylist.php
@@ -87,7 +87,7 @@
                         </a>
                     </td>
                     <td><?php echo $entry['difficulty']; ?></td>
-                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", $entry['best_time']/1000.0); ?></td>
+                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", (int)($entry['best_time']/1000.0)); ?></td>
                     <td><?php echo $entry['last_changed']; ?></td>
                 </tr>
             <?php endforeach; ?>
