+++ Source/GridDominance.Server/admin/levelscores.php
@@ -38,7 +38,7 @@
                         <?php echo fmtLevelID($entry['levelid']); ?>
                     </td>
                     <td><?php echo $entry['difficulty']; ?></td>
-                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", $entry['best_time']/1000.0); ?></td>
+                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", (int)($entry['best_time']/1000.0)); ?></td>
                     <td><?php echo $entry['last_changed']; ?></td>
                 </tr>
             <?php endforeach; ?>
@@ -65,7 +65,7 @@
 						<?php echo fmtLevelID($entry['levelid']); ?>
                     </td>
                     <td><?php echo $entry['difficulty']; ?></td>
-                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", $entry['best_time']/1000.0); ?></td>
+                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", (int)($entry['best_time']/1000.0)); ?></td>
                     <td><?php echo $entry['last_changed']; ?></td>
                 </tr>
 			<?php endforeach; ?>
@@ -92,7 +92,7 @@
 						<?php echo fmtLevelID($entry['levelid']); ?>
                     </td>
                     <td><?php echo $entry['difficulty']; ?></td>
-                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", $entry['best_time']/1000.0); ?></td>
+                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", (int)($entry['best_time']/1000.0)); ?></td>
                     <td><?php echo $entry['last_changed']; ?></td>
                 </tr>
 			<?php endforeach; ?>
@@ -119,7 +119,7 @@
 						<?php echo fmtLevelID($entry['levelid']); ?>
                     </td>
                     <td><?php echo $entry['difficulty']; ?></td>
-                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", $entry['best_time']/1000.0); ?></td>
+                    <td title="<?php echo $entry['best_time']; ?>ms" ><?php echo gmdate("H:i:s", (int)($entry['best_time']/1000.0)); ?></td>
                     <td><?php echo $entry['last_changed']; ?></td>
                 </tr>
 			<?php endforeach; ?>
